+++ Control/Joint/Abilities.hs
@@ -2,5 +2,6 @@ module Control.Joint.Abilities (module Exports) where
 
 import Control.Joint.Abilities.Liftable as Exports
 import Control.Joint.Abilities.Modulator as Exports
+import Control.Joint.Abilities.Adaptable as Exports
 import Control.Joint.Abilities.Transformer as Exports
 import Control.Joint.Abilities.Composition as Exports